+++ RELEASE_NOTES.md
@@ -167,6 +167,7 @@ To identify what has changed in between beta builds, please refer to the [bitbuc
 - ASCOM connection that is lost without any raised error will now be tried to be reconnected one time. If an error happens due to that the application will disconnect like before.
 - ASCOM get and set methods use a unified logic to ensure same behavior for all devices
 - *FLI*: Background flush is now disabled prior to readout to prevent a hung readout
+- *FLI*: Filter wheel driver now removes any extraneous `/` character from single-filter positions on CenterLine filter wheels
 - *MGEN2* now supports unattended guide star selection and calibration, and automatic meridian flips
 - Altair, MallinCam, Omegon, RisingCam and ToupTek subsampling is now available